// src/app/models/part.ts
export interface Part {
    id: string;
    partNumber: string;
    description: string;
    partColorCode: string;
    custumerPartNumber: string;
    customerColorCode: string;
    project: string;
    partGroup: string;
    partType: string;
    side: string;
    color: string;
    locationIdentifier: string;
  }